api/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
google-generativeai
SQLAlchemy
python-dotenv
markdown
